+++ targets/stm32l432/src/nfc.h
@@ -1,0 +1,7 @@
+#ifndef _NFC_H_
+#define _NFC_H_
+
+void nfc_loop();
+void nfc_init();
+
+#endif
